# server/docs/swagger.yaml
basePath: /
definitions:
...
  dtos.CreateFriendshipDTO:
    properties:
      friendID:
        type: string
      id:
        type: string
    type: object
...
  dtos.FriendshipDTO:
    properties:
      friendID:
        type: string
      friendUsername:
        type: string
      id:
        type: string
      status:
        type: string
    type: object
...
  dtos.SubjectDTO:
    properties:
      id:
        type: string
      name:
        minLength: 3
        type: string
    required:
    - name
    type: object
  dtos.UpdateUserDTO:
    properties:
      address:
        type: string
      bio:
        type: string
      birthdate:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      picture:
        type: string
      reportNumber:
        type: integer
      role:
        $ref: '#/definitions/user.Role'
      username:
        type: string
    required:
    - email
    - firstname
    - lastname
    - username
    type: object
  dtos.UserDTO:
    properties:
      activated:
        type: boolean
      address:
        type: string
      bio:
        type: string
      birthDate:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      picture:
        type: string
      reportNumber:
        type: integer
      role:
        allOf:
        - $ref: '#/definitions/user.Role'
        enum:
        - SUPERADMIN
        - ADMIN
        - USER
      username:
        minLength: 3
        type: string
    required:
    - email
    - id
    - role
    - username
    type: object
  dtos.ValidateUserDTO:
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
  dtos.VerifyCodeDTO:
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
...
  user.Role:
    enum:
    - USER
    - SUPER ADMIN
    - ADMIN
    - USER
    type: string
    x-enum-varnames:
    - DefaultRole
    - RoleSUPERADMIN
    - RoleADMIN
    - RoleUSER
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the EduMeet project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: EduMeet API Documentation
  version: "1.0"
paths:
...
  /me:
    get:
      consumes:
      - application/json
      description: Retrieve the information of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Current user information'
          schema:
            $ref: '#/definitions/dtos.UserDTO'
        "401":
          description: 'Unauthorized: Invalid token or user not found'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get current user information
      tags:
      - User
...
  /user/{id}:
    put:
      consumes:
      - application/json
      description: Update the information of a user based on their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User Information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: User information updated'
          schema:
            $ref: '#/definitions/dtos.UserDTO'
        "400":
          description: 'Bad Request: Invalid ID or data'
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: 'Unprocessable Entity: Validation errors'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'Internal Server Error: Error updating user'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user information
      tags:
      - User
  /user/friendship:
    post:
      consumes:
      - application/json
      description: Create a friendship between users
      parameters:
      - description: Friendship details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateFriendshipDTO'
      produces:
      - application/json
      responses:
        "201":
          description: 'Success: Friendship created'
          schema:
            $ref: '#/definitions/dtos.FriendshipDTO'
        "400":
          description: 'Bad Request: Invalid friendship data'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create friendship
      tags:
      - User
  /user/friendship/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a friendship between users
      parameters:
      - description: Friendship ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: 'Success: Friendship deleted'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Bad Request: Invalid friendship ID'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete friendship
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Accept a pending friendship request
      parameters:
      - description: Friendship ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Friendship accepted'
          schema:
            $ref: '#/definitions/dtos.FriendshipDTO'
        "400":
          description: 'Bad Request: Invalid friendship ID'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Accept friendship
      tags:
      - User
  /user/friendships:
    get:
      consumes:
      - application/json
      description: Retrieve a list of the user's friendships
      parameters:
      - description: Friendship status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: List of friendships'
          schema:
            items:
              $ref: '#/definitions/dtos.FriendshipDTO'
            type: array
        "500":
          description: 'Internal Server Error: Error retrieving friendships'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get friendships
      tags:
      - User
  /user/information/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve the information of a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: User information'
          schema:
            $ref: '#/definitions/dtos.UserDTO'
        "400":
          description: 'Bad Request: Invalid ID format'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'Not Found: User not found'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user information by ID
      tags:
      - User
  /user/subjects:
    get:
      consumes:
      - application/json
      description: Retrieve subjects associated with the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: List of user subjects'
          schema:
            items:
              $ref: '#/definitions/dtos.SubjectDTO'
            type: array
        "500":
          description: 'Internal Server Error: Error retrieving subjects'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user subjects
      tags:
      - User
  /user/subjects/update:
    put:
      consumes:
      - application/json
      description: Update the subjects associated with the currently authenticated
        user
      parameters:
      - description: List of subjects
        in: body
        name: body
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Subjects updated'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Bad Request: Invalid subjects'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Internal Server Error: Error updating subjects'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user subjects
      tags:
      - User
  /user/validate-user:
    post:
      consumes:
      - application/json
      description: Validate user activation code for account activation
      parameters:
      - description: Activation Code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.ValidateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: User activated successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Bad Request: Invalid activation code'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Internal Server Error: Error processing validation'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Validate user activation code
      tags:
      - User
  /user/verify:
    post:
      consumes:
      - application/json
      description: Verify the user email using the provided verification code
      parameters:
      - description: Verification Code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.VerifyCodeDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Id verified successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Bad Request: Invalid verification code'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Internal Server Error: Error processing verification'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify user email code
      tags:
      - User
swagger: "2.0"